Replace SWT browser view control with JxBrowser implementation.
We keep the SWT browser view implementation and introduce BrowserView interface
with two implementations: the old one based on the standard Eclipse's SWT browser
view control and the new one based on JxBrowser. It allows developers to switch
between two implementations for debugging purposes if it's necessary.
JxBrowser has been added to the list of project dependencies.